src/main/python/requirements.txt
numpy
torch
torchvision
Pillow
flask
matplotlib
jupyter